Simplifying unbox logic

+++ gitbox/unbox.py
@@ -150,26 +150,25 @@ def create_virtualenv(conf, dest, virtualenv_cmd, parent_virtualenv, is_dep):
         # Create virtualenv
         if conf.get('env'):
             # If installing as a dependency, link to prior virtualenv
-            if parent_virtualenv is None:
-                parent_virtualenv = conf['env']['path']
-            elif not os.path.exists(conf['env']['path']):
+            if parent_virtualenv is not None \
+                    and not os.path.exists(conf['env']['path']):
                 os.symlink(parent_virtualenv, conf['env']['path'])
 
-            # If parent is defined, try to link to the parent's virtualenv
-            if not is_dep and conf.get('parent'):
-                parent = os.path.join(os.pardir, conf['parent'])
-                if os.path.exists(parent):
-                    parent_conf = load_conf(parent)
-                    parent_venv = os.path.join(parent,
-                                               parent_conf['env']['path'])
-                    if os.path.exists(parent_venv):
-                        os.symlink(parent_venv, conf['env']['path'])
-
             if not os.path.exists(conf['env']['path']):
-                print "Creating virtualenv"
-                cmd = ([virtualenv_cmd] + conf['env']['args'] +
-                       [conf['env']['path']])
-                subprocess.check_call(cmd)
+                # If parent is defined, try to link to the parent's virtualenv
+                if not is_dep and conf.get('parent'):
+                    parent = os.path.join(os.pardir, conf['parent'])
+                    if os.path.exists(parent):
+                        parent_conf = load_conf(parent)
+                        parent_venv = os.path.join(parent,
+                                                   parent_conf['env']['path'])
+                        if os.path.exists(parent_venv):
+                            os.symlink(parent_venv, conf['env']['path'])
+                else:
+                    print "Creating virtualenv"
+                    cmd = ([virtualenv_cmd] + conf['env']['args'] +
+                           [conf['env']['path']])
+                    subprocess.check_call(cmd)
 
 
 def install(conf, dest):
